# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Todolisto
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths
basePath: /
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
#securityDefinitions:
 # Bearer:
  #  type: apiKey
   # name: Authorization
    #in: header
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /tareas:
    # our controller name
    x-swagger-router-controller: tarea
    get:
      description: Obtener la lista de tareas
      operationId: getTareas
      # define the type of response for Success "200" and Error
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Tarea"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Errors"
    post:
      description: Añadir una nueva tarea
      operationId: postTarea
      # movie info to be stored
      parameters:
        - name: data
          description: Atributos de tarea
          in: body
          required: true
          schema:
            $ref: "#/definitions/RequestTarea"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ResponseTarea"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Errors"
 
  
# complex objects have schema definitions
definitions:

  Tarea:
    type: object
    properties:
      titulo:
        type: string
        description: Nombre de la tarea a realizar
      descripcion:
        type: string
        description: descripcion
      fecha_inicio:
        type: string
        format: date-time
        description: fecha en la cual se crea la tarea.
      fecha_termino:
        type: string
        format: date-time
        description: fecha en la cual termina la tarea
      estado:
        type: number
        description: id del estado
      nombre_estado:
        type: string
        description: nombre del estado de la tarea
        
  Asignatura:
    type: object
    properties:
      nombre:
        type: string
        description: Nombre completo de la asignatura
      enfasis:
        type: array
        items:
          type: string
        description: Descripcion de lo esperado para la asignatura
      enfoque:
        type: array
        items:
          type: string
        description: Nociones y consideraciones básicas
      propositoFormativo:
        type: string
        description: Completar!!!!!!!!
  RecursoEducativo:
    type: object
    properties:
      titulo:
        type: string
        description: Titulo del recurso educativo
      descripcion:
        type: string
        description: Descripcion del recurso educativo
      asignatura:
        type: string
        description: Nombre de la asignatura a la que alude el recurso educativo
      nivel:
        $ref: "#/definitions/Nivel"
      tipoRecurso:
        type: string
        description: Tipo de recurso educativo
      links:
        type: array
        items:
          type: string
          description: Enlaces a las partes que componen el recurso educativo
  Organizacion:
    type: object
    properties:
      nombre:
        type: string
        description: Nombre completo de la organizacion
      rbd:
        type: number
        description: Numero Rol Base de Datos de la organizacion
      descripcion:
        type: string
        description: Descripcion de la organizacion
      reconocimientoOficial:
        type: string
        description: Fecha de reconocimiento oficial de la organizacion
      dependencia:
        type: string
        description: Tipo de organizacion (Municipal, Subvencionado, etc)
      sostenedor:
        type: array
        items:
          $ref: "#/definitions/User"
      orientacionReligiosa:
        type: string
        description: Nombre de la orientacion religiosa de la organizacion
      correos:
        type: array
        items:
          type: string
      telefonos:
        type: array
        items:
          type: object
          properties:
            nombre:
              type: string
              description: Nombre de la persona o lugar donde pertenece
            numero:
              type: string
              description: Numero de telefono
      web:
        type: string
        description: Pagina web de la organizacion
      director:
        type: array
        items:
          $ref: "#/definitions/User"
      mensualidad:
        type: number
        description: Costo de la mensualidad de la organizacion
      puntajeSimce:
        type: number
        description: Ultimo puntaje SIMCE obtenido por los alumnos en la organizacion
      fechaPostulacion:
        type: object
        description: Objeto que contiene la fecha de inicio y cierre de la organizacion
        properties:
          inicio:
            type: string
            format: date-time
            description: Fecha de inicio para las postulaciones de la organizacion
          cierre:
            type: string
            format: date-time
            description: Fecha de cierre para las postulaciones de la organizacion
      direccion:
        type: object
        description: Objeto que contiene la ubicacion de la organizacion
        properties:
          calle:
            type: string
            description: Nombre de la calle en donde se ubica la organizacion
          region:
            type: string
            description: Nombre de la region en donde se ubica la organizacion
          comuna:
            type: string
            description: Nombre de la comuna en donde se ubica la organizacion
      numVacantes:
        type: number
        description: Numero de vacantes disponibles de la organizacion
      totalAlumnosMatriculados:
        type: number
        description: Cantidad total de alumnos matriculados en la organizacion
      promAlumnosCurso:
        type: number
        description: Promedio general de alumnos por curso en la organizacion
      cursos:
        type: array
        items:
          $ref: "#/definitions/Curso"
      profesores:
        type: array
        items:
          $ref: "#/definitions/User"
      alumnos:
        type: array
        items:
          $ref: "#/definitions/User"
      createAt:
        type: string
        format: date-time
  User:
    type: object
    properties:
      nombre:
        type: string
        description: Nombre completo del usuario
      rut:
        type: string
        description: RUT del usuario
      password:
        type: string
        description: Clave del usuario
      correo:
        type: string
        description: Correo del usuario
      role:
        type: array
        items:
          type: string
        description: Tipo de usuario
  Nivel:
    type: object
    properties:
      sigla:
        type: string
        description: Nombre completo del usuario
      tipo_nivel:
        type: string
        description: RUT del usuario
      grado:
        type: number
        description: Correo del usuario
      descripcion:
        type: string
        description: descripción del nivel
      decreto:
        type: string
        description: Decreto para el nivel académico
      oa:
        type: array
        items:
           $ref: "#/definitions/ObjAprendizaje"
      asignatura:
        $ref: "#/definitions/Asignatura"


  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      message:
        type: string
        description: a short comment
    required:
      - success
      - message
  GetUserResponse:
    required:
      - id
      - name
      - rut
    properties:
      id:
        type: string
      name:
        type: string
      rut:
        type: string
      mail:
        type: string
  Curso:
    type: object
    properties:
      nombre:
        type: string
        description: Nombre del curso
      salaCurso:
        type: string
        description: Sala del curso
      nivel:
        $ref: "#/definitions/Nivel"
      detalles:
        type: string
        description: Detalles del curso
      planificacion:
        type: array
        description: Planificación del curso
        items:
          type: object
          properties:
            oa:
              $ref: "#/definitions/ObjAprendizaje"
            detalles:
              type: string
            recursos:
              type: array
              items:
                $ref: "#/definitions/RecursoEducativo"
      profesores:
        type: array
        items:
          $ref: "#/definitions/User"
      alumnos:
        type: array
        items:
          $ref: "#/definitions/User"
  Actividad:
    type: object
    properties:
      autor:
        $ref: "#/definitions/User"
      dificultad:
        type: string
        description: Nivel de dificultad de la actividad ('Alta' - 'Media' - 'Baja')
      nivelAprendizaje:
        type: string
        description: Nivel de aprendizaje ('Recordar', 'Comprender', 'Aplicar','Analizar','Evaluar', 'Crear')
      tipoPregunta:
        type: string
        description: Tipo de pregunta ('Enumeracion', 'Conocimiento especifico', 'Ensayo','Verdadero y Falso','Alternativas')
      nivel:
        $ref: "#/definitions/Nivel"
      tema:
        type: string
        description: Nombre del tema de la actividad
      subTema:
        type: string
        description: Nombre del subtema de la actividad
      preguntaEnunciado:
        type: string
        description: Enunciado de la pregunta
      alternativas:
        type: array
        items:
          type: string
        description: Alternativas correspondientes al enunciado de la pregunta
      respuesta:
        type: string
        description: Respuesta de referencia
      respuestaVerdaderoFalso:
        type: boolean
        description: Respuesta de referencia para pregunta de Verdadero y Falso (true - false)
      puntajeTotal:
        type: number
        description: Puntaje total de la pregunta
      retroalimentacion:
        type: array
        items:
          type: string
        description: Enlaces y/o información complementaria para ser entregada como retroalimentación
  Evaluacion:
    type: object
    properties:
      titulo:
        type: string
        description: Titulo de la evaluacion
      profesorAutor:
        $ref: "#/definitions/User"
      nivelAprendizaje:
        type: string
        description: Nivel de aprendizaje ('Recordar', 'Comprender', 'Aplicar','Analizar','Evaluar', 'Crear')
      asignatura:
        $ref: "#/definitions/Asignatura"
      detalles:
        type: string
        description: Detalles que explican de qué trata esta evaluación
      tipoEjecucion:
        type: string
        description: Formas en las que se llevará acabo la evaluación
      retroalimentacion:
        $ref: "#/definitions/RecursoEducativo"
      actividades:
        type: array
        items:
          $ref: "#/definitions/Actividad"
  Orientacion:
    type: object
    properties:
      introduccion:
        type: string
        description: Introduccion
      programaEstudio:
        type: string
        description: Formas en las que se llevará acabo la evaluación
      asignaturas:
        type: array
        items:
          $ref: "#/definitions/Nivel"
  Errors:
    type: object
    properties:
      errors:
        type: array
        items:
          type: object
          properties:
            titulo:
              type: string
            detalle:
              type: string
            link:
              type: string
            estado:
              type: string
  Response:
    type: object
    properties:
      link:
        type: string
  ResponseNivel:
    type: object
    properties:
      link:
        type: string
      data:
        type: array
        items:
          $ref: "#/definitions/Nivel"
      type:
        type: string
        
  ResponseTarea:
    type: object
    properties:
      link:
        type: string
      data:
        type: array
        items:
          $ref: "#/definitions/Tarea"
      type:
        type: string
  ResponseOrientacion:
    type: object
    properties:
      link:
        type: string
      data:
        type: array
        items:
          $ref: "#/definitions/Orientacion"
      type:
        type: string
  PlanEstudio:
    type: object
    properties:
      nombre:
        type: string
        description: Nombre del plan de estudio
      asignaturas:
        type: array
        items:
          type: object
          properties:
            nivel:
              $ref: "#/definitions/Nivel"
            horasMensuales:
              type: number
              description: Cantidad de horas mensuales
            horasAnuales:
              type: number
              description: Cantidad de horas anuales
      tipo:
        type: string
        description: Tipo de jornada ('Con JEC', 'Sin JEC')
      horasLibreDisposicion:
        type: number
        description: Cantidad de horas de libre disposicion
      totalTiempoMinFG:
        type: number
        description: Tiempo total de Formacion General
      totalTiempoMinFD:
        type: number
        description: Tiempo total de Formacion Diferenciada
# Objetivos de Aprendizaje
  ObjAprendizaje:
    type: object
    properties:
      identificador:
        type: string
        description: numero del objetivo de aprendizaje
      tipo:
        type: string
        description: Tipo de aprendizaje ('Habilidad', 'Conocimiento', 'Actitud')
      descripcion:
        type: string
        description: descripcion de objetivos de aprendizaje
  ResponseObjAprendizaje:
    type: object
    properties:
      link:
        type: string
      data:
        type: array
        items:
          $ref: "#/definitions/ObjAprendizaje"
      type:
        type: string
  ResponseAsignatura:
    type: object
    properties:
      link:
        type: string
      data:
        type: array
        items:
          type: string
          description: Enlaces a las partes que componen el recurso educativo
  ResponseOrganizacion:
    type: object
    properties:
      link:
        type: string
      data:
        type: array
        items:
          $ref: "#/definitions/Organizacion"
      type:
        type: string
  ResponsePlanEstudio:
    type: object
    properties:
      link:
        type: string
      data:
        type: array
        items:
          $ref: "#/definitions/PlanEstudio"
      type:
        type: string
  ResponseEvaluacion:
    type: object
    properties:
      link:
        type: string
      data:
        type: array
        items:
          $ref: "#/definitions/Evaluacion"
      type:
        type: string
  ResponseRecursoEducativo:
    type: object
    properties:
      link:
        type: string
      data:
        type: array
        items:
          $ref: "#/definitions/RecursoEducativo"
      type:
        type: string
  ResponseActividad:
    type: object
    properties:
      link:
        type: string
      data:
        type: array
        items:
          $ref: "#/definitions/Actividad"
      type:
        type: string
  ResponseCurso:
    type: object
    properties:
      link:
        type: string
      data:
        type: array
        items:
          $ref: "#/definitions/Curso"
      type:
        type: string
  RequestNivel:
    type: object
    properties:
      sigla:
        type: string
        description: Nombre completo del usuario
      tipo_nivel:
        type: string
        description: RUT del usuario
      grado:
        type: number
        description: Correo del usuario
      descripcion:
        type: string
        description: descripción del nivel
      decreto:
        type: string
        description: Decreto para el nivel académico
      oa:
        type: array
        items:
           type: string
      asignatura:
        type: string
        
        
  RequestTarea:
    type: object
    properties:
      titulo:
        type: string
        description: Nombre completo del usuario
      descripcion:
        type: string
        description: RUT del usuario
      fecha_inicio:
        type: string
        description: Correo del usuario
      fecha_termino:
        type: string
        description: descripción del nivel
      estado:
        type: number
        description: Decreto para el nivel académico
      nombre_estado:
        type: string

  RequestOrientacion:
    type: object
    properties:
      introduccion:
        type: string
        description: Introduccion
      programaEstudio:
        type: string
        description: Formas en las que se llevará acabo la evaluación
      asignaturas:
        type: array
        items:
          type: string
  RequestOrganizacion:
    type: object
    properties:
      nombre:
        type: string
        description: Nombre completo de la organizacion
      rbd:
        type: number
        description: Numero Rol Base de Datos de la organizacion
      descripcion:
        type: string
        description: Descripcion de la organizacion
      reconocimientoOficial:
        type: string
        description: Fecha de reconocimiento oficial de la organizacion
      dependencia:
        type: string
        description: Tipo de organizacion (Municipal, Subvencionado, etc)
      sostenedor:
        type: array
        items:
          type: string
      orientacionReligiosa:
        type: string
        description: Nombre de la orientacion religiosa de la organizacion
      correos:
        type: array
        items:
          type: string
      telefonos:
        type: array
        items:
          type: object
          properties:
            nombre:
              type: string
              description: Nombre de la persona o lugar donde pertenece
            numero:
              type: string
              description: Numero de telefono
      web:
        type: string
        description: Pagina web de la organizacion
      director:
        type: array
        items:
          type: string
      mensualidad:
        type: number
        description: Costo de la mensualidad de la organizacion
      puntajeSimce:
        type: number
        description: Ultimo puntaje SIMCE obtenido por los alumnos en la organizacion
      fechaPostulacion:
        type: object
        description: Objeto que contiene la fecha de inicio y cierre de la organizacion
        properties:
          inicio:
            type: string
            format: date
            description: Fecha de inicio para las postulaciones de la organizacion
          cierre:
            type: string
            format: date
            description: Fecha de cierre para las postulaciones de la organizacion
      direccion:
        type: object
        description: Objeto que contiene la ubicacion de la organizacion
        properties:
          calle:
            type: string
            description: Nombre de la calle en donde se ubica la organizacion
          region:
            type: string
            description: Nombre de la region en donde se ubica la organizacion
          comuna:
            type: string
            description: Nombre de la comuna en donde se ubica la organizacion
      numVacantes:
        type: number
        description: Numero de vacantes disponibles de la organizacion
      totalAlumnosMatriculados:
        type: number
        description: Cantidad total de alumnos matriculados en la organizacion
      promAlumnosCurso:
        type: number
        description: Promedio general de alumnos por curso en la organizacion
      cursos:
        type: array
        items:
          type: string
      profesores:
        type: array
        items:
          type: string
      alumnos:
        type: array
        items:
          type: string
      createAt:
        type: string
        format: date
  RequestEvaluacion:
    type: object
    properties:
      titulo:
        type: string
        description: Titulo de la evaluacion
      profesorAutor:
        type: string
      nivelAprendizaje:
        type: string
        description: Nivel de aprendizaje ('Recordar', 'Comprender', 'Aplicar','Analizar','Evaluar', 'Crear')
      asignatura:
        type: string
      detalles:
        type: string
        description: Detalles que explican de qué trata esta evaluación
      tipoEjecucion:
        type: string
        description: Formas en las que se llevará acabo la evaluación
      retroalimentacion:
        type: string
      actividades:
        type: array
        items:
          type: string
  RequestPlanEstudio:
    type: object
    properties:
      nombre:
        type: string
        description: Nombre del plan de estudio
      asignaturas:
        type: array
        items:
          type: object
          properties:
            nivel:
              type: string
            horasMensuales:
              type: number
              description: Cantidad de horas mensuales
            horasAnuales:
              type: number
              description: Cantidad de horas anuales
      tipo:
        type: string
        description: Tipo de jornada ('Con JEC', 'Sin JEC')
      horasLibreDisposicion:
        type: number
        description: Cantidad de horas de libre disposicion
      totalTiempoMinFG:
        type: number
        description: Tiempo total de Formacion General
      totalTiempoMinFD:
        type: number
        description: Tiempo total de Formacion Diferenciada
      createAt:
        type: string
        format: date-time
  RequestCurso:
    type: object
    properties:
      nombre:
        type: string
        description: Nombre del curso
      salaCurso:
        type: string
        description: Sala del curso
      nivel:
        type: string
        description: Nivel al que corresponde (ID)
      detalles:
        type: string
        description: Detalles del curso
      planificacion:
        type: array
        description: Planificación del curso
        items:
          type: object
          properties:
            oa:
              type: string
              description: Objetivo de Aprendizaje (ID)
            detalles:
              type: string
            recursos:
              type: array
              items:
                type: string
                description: Recurso educativo (ID)
      profesores:
        type: array
        items:
          type: string
          description: Alumno (ID)
      alumnos:
        type: array
        items:
          type: string
          description: Profesor (ID)
  RequestActividad:
    type: object
    properties:
      autor:
        type: string
        description: Autor de la actividad (ID)
      dificultad:
        type: string
        description: Nivel de dificultad de la actividad ('Alta' - 'Media' - 'Baja')
      nivelAprendizaje:
        type: string
        description: Nivel de aprendizaje ('Recordar', 'Comprender', 'Aplicar','Analizar','Evaluar', 'Crear')
      tipoPregunta:
        type: string
        description: Tipo de pregunta ('Enumeracion', 'Conocimiento especifico', 'Ensayo','Verdadero y Falso','Alternativas')
      nivel:
        type: string
        description: Nivel al que corresponde (ID)
      tema:
        type: string
        description: Nombre del tema de la actividad
      subTema:
        type: string
        description: Nombre del subtema de la actividad
      preguntaEnunciado:
        type: string
        description: Enunciado de la pregunta
      alternativas:
        type: array
        items:
          type: string
        description: Alternativas correspondientes al enunciado de la pregunta
      respuesta:
        type: string
        description: Respuesta de referencia
      respuestaVerdaderoFalso:
        type: boolean
        description: Respuesta de referencia para pregunta de Verdadero y Falso (true - false)
      puntajeTotal:
        type: number
        description: Puntaje total de la pregunta
      retroalimentacion:
        type: array
        items:
          type: string
        description: Enlaces y/o información complementaria para ser entregada como retroalimentación
  RequestRecursoEducativo:
    type: object
    properties:
      titulo:
        type: string
        description: Titulo del recurso educativo
      descripcion:
        type: string
        description: Descripcion del recurso educativo
      asignatura:
        type: string
        description: Nombre de la asignatura a la que alude el recurso educativo
      nivel:
        type: string
        description: Nivel al que corresponde el recurso educativo (ID)
      tipoRecurso:
        type: string
        description: Tipo de recurso educativo
      links:
        type: array
        items:
          type: string
          description: Enlaces a las partes que componen el recurso educativo